added relations now other entities can track of the changes

+++ src/main/java/com/gri/agriconnect/service/CommentService.java
@@ -1,7 +1,11 @@
 package com.gri.agriconnect.service;
 
 import com.gri.agriconnect.model.Comment;
+import com.gri.agriconnect.model.Post;
+import com.gri.agriconnect.model.Product;
 import com.gri.agriconnect.repository.CommentRepository;
+import com.gri.agriconnect.repository.PostRepository;
+import com.gri.agriconnect.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -9,14 +13,45 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-
 @Service
 public class CommentService {
 
+    private final CommentRepository commentRepository;
+    private final PostService postService;
+    private final ProductService productService;
+
     @Autowired
-    private CommentRepository commentRepository;
+    public CommentService(CommentRepository commentRepository, PostService postService, ProductService productService) {
+        this.commentRepository = commentRepository;
+        this.postService = postService;
+        this.productService = productService;
+    }
 
     public Comment saveComment(Comment comment) {
+        boolean isValid = false;
+
+        Optional<Post> postOpt = postService.getPostById(comment.getPostId());
+        if (postOpt.isPresent()) {
+            Post post = postOpt.get();
+            Comment savedComment = commentRepository.save(comment);
+            post.getCommentIds().add(savedComment.getCommentId());
+            postService.updatePost(post.getPostId(), post);
+            isValid = true;
+        }
+
+        Optional<Product> productOpt = productService.getProductById(comment.getPostId());
+        if (productOpt.isPresent()) {
+            Product product = productOpt.get();
+            Comment savedComment = commentRepository.save(comment);
+            product.getCommentIds().add(savedComment.getCommentId());
+            productService.updateProduct(product.getProductId(), product);
+            isValid = true;
+        }
+
+        if (!isValid) {
+            throw new IllegalArgumentException("Post or Product with ID " + comment.getPostId() + " does not exist.");
+        }
+
         comment.setUpdatedAt(LocalDateTime.now());
         return commentRepository.save(comment);
     }
@@ -34,7 +69,28 @@ public class CommentService {
     }
 
     public void deleteComment(String commentId) {
-        commentRepository.deleteById(commentId);
+        Optional<Comment> commentOpt = commentRepository.findById(commentId);
+        if (commentOpt.isPresent()) {
+            Comment comment = commentOpt.get();
+
+            Optional<Post> postOpt = postService.getPostById(comment.getPostId());
+            if (postOpt.isPresent()) {
+                Post post = postOpt.get();
+                post.getCommentIds().remove(commentId);
+                postService.updatePost(post.getPostId(), post);
+            }
+
+            Optional<Product> productOpt = productService.getProductById(comment.getPostId());
+            if (productOpt.isPresent()) {
+                Product product = productOpt.get();
+                product.getCommentIds().remove(commentId);
+                productService.updateProduct(product.getProductId(), product);
+            }
+
+            commentRepository.deleteById(commentId);
+        } else {
+            throw new IllegalArgumentException("Comment with ID " + commentId + " does not exist.");
+        }
     }
 
     public Comment updateComment(String commentId, Comment updatedComment) {
@@ -51,5 +107,3 @@ public class CommentService {
         });
     }
 }
-
-
