+++ src/main/java/com/gri/agriconnect/service/MessageService.java
@@ -1,18 +1,37 @@
 package com.gri.agriconnect.service;
 
+import com.gri.agriconnect.model.Conversation;
 import com.gri.agriconnect.model.Message;
 import com.gri.agriconnect.repository.MessageRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class MessageService {
+
+    private final MessageRepository messageRepository;
+    private final ConversationService conversationService;
+
     @Autowired
-    private MessageRepository messageRepository;
+    public MessageService(MessageRepository messageRepository, ConversationService conversationService) {
+        this.messageRepository = messageRepository;
+        this.conversationService = conversationService;
+    }
 
     public Message saveMessage(Message message) {
-        return messageRepository.save(message);
+        Optional<Conversation> conversationOpt = conversationService.getConversationById(message.getConversationId());
+        if (conversationOpt.isPresent()) {
+            Conversation conversation = conversationOpt.get();
+            Message savedMessage = messageRepository.save(message);
+            conversation.getMessageIds().add(savedMessage.getMessageId());
+            conversationService.updateConversation(conversation.getConversationId(), conversation);
+            return savedMessage;
+        } else {
+            throw new IllegalArgumentException("Conversation with ID " + message.getConversationId() + " does not exist.");
+        }
     }
 
     public List<Message> getMessagesByConversationId(String conversationId) {
@@ -24,7 +43,19 @@ public class MessageService {
     }
 
     public void deleteMessage(String messageId) {
-        messageRepository.deleteById(messageId);
+        Optional<Message> messageOpt = messageRepository.findById(messageId);
+        if (messageOpt.isPresent()) {
+            Message message = messageOpt.get();
+            Optional<Conversation> conversationOpt = conversationService.getConversationById(message.getConversationId());
+            if (conversationOpt.isPresent()) {
+                Conversation conversation = conversationOpt.get();
+                conversation.getMessageIds().remove(messageId);
+                conversationService.updateConversation(conversation.getConversationId(), conversation);
+            }
+            messageRepository.deleteById(messageId);
+        } else {
+            throw new IllegalArgumentException("Message with ID " + messageId + " does not exist.");
+        }
     }
 
     public Message updateMessage(String messageId, Message updatedMessage) {
@@ -38,3 +69,4 @@ public class MessageService {
         });
     }
 }
+
