+++ src/main/java/com/gri/agriconnect/service/PostService.java
@@ -1,9 +1,11 @@
 package com.gri.agriconnect.service;
 
 import com.gri.agriconnect.model.Post;
+import com.gri.agriconnect.model.User;
 import com.gri.agriconnect.repository.PostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -11,12 +13,24 @@ import java.util.Optional;
 @Service
 public class PostService {
 
+    private final PostRepository postRepository;
+    private final UserService userService;
+
     @Autowired
-    private PostRepository postRepository;
+    public PostService(PostRepository postRepository, UserService userService) {
+        this.postRepository = postRepository;
+        this.userService = userService;
+    }
 
     public Post savePost(Post post) {
-        post.setUpdatedAt(LocalDateTime.now());
-        return postRepository.save(post);
+        Optional<User> userOpt = userService.getUserById(post.getUserId());
+        if (userOpt.isPresent()) {
+            Post savedPost = postRepository.save(post);
+            userService.addPostToUser(post.getUserId(), savedPost.getPostId());
+            return savedPost;
+        } else {
+            throw new IllegalArgumentException("User with ID " + post.getUserId() + " does not exist.");
+        }
     }
 
     public List<Post> getAllPosts() {
@@ -32,7 +46,14 @@ public class PostService {
     }
 
     public void deletePost(String postId) {
-        postRepository.deleteById(postId);
+        Optional<Post> postOpt = postRepository.findById(postId);
+        if (postOpt.isPresent()) {
+            Post post = postOpt.get();
+            userService.removePostFromUser(post.getUserId(), postId);
+            postRepository.deleteById(postId);
+        } else {
+            throw new IllegalArgumentException("Post with ID " + postId + " does not exist.");
+        }
     }
 
     public Post updatePost(String postId, Post updatedPost) {
@@ -42,7 +63,8 @@ public class PostService {
             post.setFavoriteCount(updatedPost.getFavoriteCount());
             post.setLikeCount(updatedPost.getLikeCount());
             post.setCommentCount(updatedPost.getCommentCount());
-            post.setCommentIds(updatedPost.getCommentIds());
+            post.setCategoryTags(updatedPost.getCategoryTags());
+            post.setImageLinks(updatedPost.getImageLinks());
             post.setUpdatedAt(LocalDateTime.now());
             return postRepository.save(post);
         }).orElseGet(() -> {
@@ -53,4 +75,3 @@ public class PostService {
         });
     }
 }
-
